# Optimise allocations in ParseAccept