--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/munnerz/goautoneg
+
+go 1.18