--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/munnerz/goautoneg
 
 go 1.18
+
+require golang.org/x/exp v0.0.0-20230224173230-c95f2b4c22f2