This PR contains two small optimisations that overall produce the following improvements to `ParseAccept`:

```
goos: darwin
goarch: arm64
pkg: github.com/munnerz/goautoneg
                                                                                                          │ original.txt │              new.txt               │
                                                                                                          │    sec/op    │   sec/op     vs base               │
ParseAccept/#00-10                                                                                          24.445n ± 0%   6.540n ± 1%  -73.25% (p=0.002 n=6)
ParseAccept/application/json-10                                                                              90.67n ± 1%   66.41n ± 1%  -26.75% (p=0.002 n=6)
ParseAccept/application/json,text/plain-10                                                                   152.7n ± 1%   123.7n ± 1%  -18.99% (p=0.002 n=6)
ParseAccept/application/json;q=0.9,text/plain-10                                                             214.5n ± 0%   164.5n ± 1%  -23.29% (p=0.002 n=6)
ParseAccept/application/xml,application/xhtml+xml,text/html;q=0.9,text/plain;q=0.8,image/png,*/*;q=0.5-10    636.6n ± 1%   528.9n ± 0%  -16.92% (p=0.002 n=6)

                                                                                                          │ original.txt │                new.txt                 │
                                                                                                          │     B/op     │    B/op     vs base                    │
ParseAccept/#00-10                                                                                            24.00 ± 0%    0.00 ± 0%  -100.00% (p=0.002 n=6)
ParseAccept/application/json-10                                                                               72.00 ± 0%   48.00 ± 0%   -33.33% (p=0.002 n=6)
ParseAccept/application/json,text/plain-10                                                                   120.00 ± 0%   80.00 ± 0%   -33.33% (p=0.002 n=6)
ParseAccept/application/json;q=0.9,text/plain-10                                                             168.00 ± 0%   80.00 ± 0%   -52.38% (p=0.002 n=6)
ParseAccept/application/xml,application/xhtml+xml,text/html;q=0.9,text/plain;q=0.8,image/png,*/*;q=0.5-10     456.0 ± 0%   240.0 ± 0%   -47.37% (p=0.002 n=6)

                                                                                                          │ original.txt │                new.txt                 │
                                                                                                          │  allocs/op   │ allocs/op   vs base                    │
ParseAccept/#00-10                                                                                            1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
ParseAccept/application/json-10                                                                               2.000 ± 0%   1.000 ± 0%   -50.00% (p=0.002 n=6)
ParseAccept/application/json,text/plain-10                                                                    2.000 ± 0%   1.000 ± 0%   -50.00% (p=0.002 n=6)
ParseAccept/application/json;q=0.9,text/plain-10                                                              3.000 ± 0%   1.000 ± 0%   -66.67% (p=0.002 n=6)
ParseAccept/application/xml,application/xhtml+xml,text/html;q=0.9,text/plain;q=0.8,image/png,*/*;q=0.5-10     5.000 ± 0%   1.000 ± 0%   -80.00% (p=0.002 n=6)

```

The first optimisation (9895b45) uses `golang.org/x/exp/slices.SortFunc` instead of `sort.Sort`. This requires Go 1.18+ as it uses generics.

The second optimisation (7248a2f) drops parsing of parameters. According to [the spec](https://httpwg.org/specs/rfc9110.html#field.accept), this is no longer widely used. Dropping this parsing has the benefit of not needing to allocate for the `Accept.Params` field.