@pracucci 2024-06-07
This change breaks https://github.com/prometheus/common which is using `Params` and vendoring this library since https://github.com/prometheus/common/pull/625