@pracucci 2024-06-07
> This change breaks https://github.com/prometheus/common which is using `Params` and vendoring this library since [prometheus/common#625](https://github.com/prometheus/common/pull/625)

See https://github.com/munnerz/goautoneg/pull/6 for the same PR, but re-introducing `Params` (so only 1 out of the 2 optimizations).